+++ persistence/persistence-deltaspike/src/main/java/uk/gov/justice/services/persistence/EntityManagerFlushInterceptorProvider.java
@@ -1,16 +1,24 @@
 package uk.gov.justice.services.persistence;
 
+import static java.util.Collections.emptyList;
 import static java.util.List.of;
 import static uk.gov.justice.services.core.annotation.Component.EVENT_LISTENER;
 
+import uk.gov.justice.services.common.configuration.errors.event.EventErrorHandlingConfiguration;
 import uk.gov.justice.services.core.interceptor.InterceptorChainEntry;
 import uk.gov.justice.services.core.interceptor.InterceptorChainEntryProvider;
 
 import java.util.List;
 
+import javax.inject.Inject;
+
 public class EntityManagerFlushInterceptorProvider implements InterceptorChainEntryProvider {
 
     private static final int PRIORITY = 220;
+
+    @Inject
+    private EventErrorHandlingConfiguration eventErrorHandlingConfiguration;
+
 
     @Override
     public String component() {
@@ -19,6 +27,11 @@ public class EntityManagerFlushInterceptorProvider implements InterceptorChainEn
 
     @Override
     public List<InterceptorChainEntry> interceptorChainTypes() {
-        return of(new InterceptorChainEntry(PRIORITY, EntityManagerFlushInterceptor.class));
+
+        if (eventErrorHandlingConfiguration.isEventErrorHandlingEnabled()) {
+            return of(new InterceptorChainEntry(PRIORITY, EntityManagerFlushInterceptor.class));
+        }
+
+        return emptyList();
     }
 }